requirements_simple.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --generate-hashes --output-file=requirements.txt --strip-extras requirements.in
#
alembic==1.18.0
# via -r requirements.in
blinker==1.9.0
# via
#   -r requirements.in
#   flask
boto3==1.42.25
# via -r requirements.in
botocore==1.42.25
# via
#   boto3
#   s3transfer
certifi==2025.11.12
# via requests
charset-normalizer==3.4.0
# via
#   reportlab
#   requests
click==8.3.1
# via
#   -r requirements.in
#   flask
flask==3.1.2
# via
#   -r requirements.in
#   flask-login
#   flask-wtf
flask-login==0.6.3
# via -r requirements.in
flask-wtf==1.2.2
# via -r requirements.in
greenlet==3.3.0
# via sqlalchemy
gunicorn==23.0.0
# via -r requirements.in
idna==3.6
# via requests
itsdangerous==2.2.0
# via
#   -r requirements.in
#   flask
#   flask-wtf
jinja2==3.1.6
# via
#   -r requirements.in
#   flask
jmespath==1.0.1
# via
#   boto3
#   botocore
mako==1.3.10
# via alembic
markupsafe==3.0.3
# via
#   -r requirements.in
#   flask
#   jinja2
#   mako
#   werkzeug
#   wtforms
packaging==25.0
# via gunicorn
pillow==12.1.0
# via
#   -r requirements.in
#   reportlab
psycopg2-binary==2.9.11
# via -r requirements.in
pymupdf==1.26.7
# via -r requirements.in
pypng==0.20220715.0
# via -r requirements.in
python-dateutil==2.9.0.post0
# via botocore
python-dotenv==1.2.1
# via -r requirements.in
python-slugify==8.0.4
# via -r requirements.in
qrcode==8.2
# via -r requirements.in
reportlab==4.4.7
# via -r requirements.in
requests==2.32.5
# via
#   -r requirements.in
#   stripe
s3transfer==0.16.0
# via boto3
six==1.17.0
# via python-dateutil
sqlalchemy==2.0.45
# via alembic
stripe==14.1.0
# via -r requirements.in
text-unidecode==1.3
# via python-slugify
typing-extensions==4.15.0
# via
#   alembic
#   sqlalchemy
#   stripe
urllib3==2.6.2
# via
#   botocore
#   requests
werkzeug==3.1.4
# via
#   -r requirements.in
#   flask
#   flask-login
wtforms==3.2.1
# via
#   -r requirements.in
#   flask-wtf
